change lat and long to bigints on the mssql db and udfs related

+++ database/mssql/it2s_db.sql
@@ -44,8 +44,8 @@ GO
 
 create table it2s_db.RSU(
     emitter_station_id int foreign key references it2s_db.Emitter(station_id) not null,
-    latitude int not null,
-    longitude int not null,
+    latitude BIGINT not null,
+    longitude BIGINT not null,
     primary key(emitter_station_id)
 )
 GO
@@ -54,8 +54,8 @@ GO
 create table it2s_db.CPM(
     rsu_station_id int foreign key references it2s_db.RSU(emitter_station_id) not null,
     event_timestamp int not null,
-    latitude int not null,
-    longitude int not null, 
+    latitude BIGINT not null,
+    longitude BIGINT not null, 
     quadtree bigint check(0 <= quadtree and quadtree <= 68719476736) not null,
     primary key(rsu_station_id, event_timestamp)
 )
@@ -66,8 +66,8 @@ create table it2s_db.PerceivedObject(
     cpm_station_id int not null, 
     event_timestamp int not null, 
 	perceived_object_id int not null,
-    latitude int not null,
-    longitude int not null, 
+    latitude BIGINT not null,
+    longitude BIGINT not null, 
     quadtree bigint check(0 <= quadtree and quadtree <= 68719476736) not null,
     x_distance int not null,
     y_distance int not null,
@@ -85,7 +85,7 @@ GO
 --     event_timestamp int not null,
 --     perceived_object_id int not null,
 --     latitude int not null,
---     longitude int not null,
+--     longitude BIGINT not null,
 --     foreign key(cpm_station_id, event_timestamp, perceived_object_id) references it2s_db.PerceivedObject( cpm_station_id, event_timestamp, perceived_object_id),
 --     primary key(cpm_station_id, event_timestamp)
 -- )
@@ -97,8 +97,8 @@ create table it2s_db.CAM(
     event_timestamp int not null,
     station_type int not null,
     speed int not null,
-    latitude int not null,
-    longitude int not null,
+    latitude BIGINT not null,
+    longitude BIGINT not null,
     quadtree bigint check(0 <= quadtree and quadtree <= 68719476736) not null, -- check limite quadtree zoom 18
     primary key(station_id, event_timestamp)
 )
@@ -109,8 +109,8 @@ create table it2s_db.VAM(
     emitter_station_id int foreign key references it2s_db.SmartPhone(emitter_station_id),
     event_timestamp int not null,
     station_type int not null,
-    latitude int not null,
-    longitude int not null,
+    latitude BIGINT not null,
+    longitude BIGINT not null,
     quadtree bigint check(0 <= quadtree and quadtree <= 68719476736) not null,
     primary key(emitter_station_id, event_timestamp)
 )
@@ -122,8 +122,8 @@ create table it2s_db.DENM(
     event_timestamp int not null,
     cause_code int not null,
     sub_cause_code int not null,
-    latitude int not null,
-    longitude int not null,
+    latitude BIGINT not null,
+    longitude BIGINT not null,
     duration int,
     quadtree bigint check(0 <= quadtree and quadtree <= 68719476736) not null,
     primary key(emitter_station_id, event_timestamp)
